Refactoring to split json and sql generation into separate classes
